Refactor service profiles resource+data_source schemas and mapping helpers
