+++ docs/resources/equinix_fabric_connection.md
@@ -550,6 +550,13 @@ resource "equinix_fabric_connection" "epl" {
 }
 ```
 
+### Notes:
+
+Port to IBM Connections could be modified from IBM Service Provider Side by using parameters passed to additional_info field:
+* `{"key": "ASN", "value": "1111"}`
+* `{"key": "Global", "value": "false"}`
+* `{"key": "BGP_IBM_CIDR", "value": "172.16.0.18/30"}`
+* `{"key": "BGP_CER_CIDR", "value": "172.16.0.19/30"}`
 
 <!-- schema generated by tfplugindocs -->
 ## Schema
@@ -571,8 +578,6 @@ resource "equinix_fabric_connection" "epl" {
 - `project` (Block Set, Max: 1) Project information (see [below for nested schema](#nestedblock--project))
 - `redundancy` (Block Set, Max: 1) Connection Redundancy Configuration (see [below for nested schema](#nestedblock--redundancy))
 - `timeouts` (Block, Optional) (see [below for nested schema](#nestedblock--timeouts))
-
--~> **NOTE:** Port to IBM Connections could be modified from IBM Service Provider Side by using parameters passed to additional_info field:  `{"key": "ASN", "value": "1111"}` `{"key": "Global", "value": "false"}` `{"key": "BGP_IBM_CIDR", "value": "172.16.0.18/30"}` `{"key": "BGP_CER_CIDR", "value": "172.16.0.19/30"}`
 
 ### Read-Only
 
@@ -755,13 +760,14 @@ Read-Only:
 
 Read-Only:
 
-- `account_name` (String) Account Name
-- `account_number` (Number) Account Number
-- `global_cust_id` (String) Global Customer organization identifier
-- `global_org_id` (String) Global organization identifier
-- `global_organization_name` (String) Global organization name
-- `org_id` (Number) Customer organization identifier
-- `organization_name` (String) Customer organization name
+- `account_name` (String) Legal name of the accountholder.
+- `account_number` (Number) Equinix-assigned account number.
+- `global_cust_id` (String) Equinix-assigned ID of the subscriber's parent organization.
+- `global_org_id` (String) Equinix-assigned ID of the subscriber's parent organization.
+- `global_organization_name` (String) Equinix-assigned name of the subscriber's parent organization.
+- `org_id` (Number) Equinix-assigned ID of the subscriber's organization.
+- `organization_name` (String) Equinix-assigned name of the subscriber's organization.
+- `ucm_id` (String) Enterprise datastore id
 
 
 
@@ -982,13 +988,14 @@ Read-Only:
 
 Read-Only:
 
-- `account_name` (String) Account Name
-- `account_number` (Number) Account Number
-- `global_cust_id` (String) Global Customer organization identifier
-- `global_org_id` (String) Global organization identifier
-- `global_organization_name` (String) Global organization name
-- `org_id` (Number) Customer organization identifier
-- `organization_name` (String) Customer organization name
+- `account_name` (String) Legal name of the accountholder.
+- `account_number` (Number) Equinix-assigned account number.
+- `global_cust_id` (String) Equinix-assigned ID of the subscriber's parent organization.
+- `global_org_id` (String) Equinix-assigned ID of the subscriber's parent organization.
+- `global_organization_name` (String) Equinix-assigned name of the subscriber's parent organization.
+- `org_id` (Number) Equinix-assigned ID of the subscriber's organization.
+- `organization_name` (String) Equinix-assigned name of the subscriber's organization.
+- `ucm_id` (String) Enterprise datastore id
 
 
 
@@ -1033,7 +1040,7 @@ Read-Only:
 
 Optional:
 
-- `group` (String) Redundancy group identifier (UUID of primary connection)
+- `group` (String) Redundancy group identifier (Use the redundancy.0.group UUID of primary connection; e.g. one(equinix_fabric_connection.primary_port_connection.redundancy).group or equinix_fabric_connection.primary_port_connection.redundancy.0.group)
 - `priority` (String) Connection priority in redundancy group - PRIMARY, SECONDARY
 
 
@@ -1060,6 +1067,7 @@ Read-Only:
 - `global_organization_name` (String)
 - `org_id` (Number)
 - `organization_name` (String)
+- `ucm_id` (String)
 
 
 <a id="nestedatt--change_log"></a>
