+++ equinix/resource_fabric_port.go
@@ -398,7 +398,7 @@ func fabricPortsListToTerra(ports v4.AllPortsResponse) []map[string]interface{} 
 			"service_type":        port.ServiceType,
 			"operation":           portOperationToTerra(port.Operation),
 			"redundancy":          PortRedundancyToTerra(port.Redundancy),
-			"account":             accountToTerra(port.Account),
+			"account":             equinix_fabric_schema.AccountToTerra(port.Account),
 			"change_log":          equinix_fabric_schema.ChangeLogToTerra(port.Changelog),
 			"location":            equinix_fabric_schema.LocationToTerra(port.Location),
 			"device":              portDeviceToTerra(port.Device),
@@ -438,7 +438,7 @@ func setFabricPortMap(d *schema.ResourceData, port v4.Port) diag.Diagnostics {
 		"service_type":        port.ServiceType,
 		"operation":           portOperationToTerra(port.Operation),
 		"redundancy":          PortRedundancyToTerra(port.Redundancy),
-		"account":             accountToTerra(port.Account),
+		"account":             equinix_fabric_schema.AccountToTerra(port.Account),
 		"change_log":          equinix_fabric_schema.ChangeLogToTerra(port.Changelog),
 		"location":            equinix_fabric_schema.LocationToTerra(port.Location),
 		"device":              portDeviceToTerra(port.Device),
