+++ equinix/resource_fabric_service_profile.go
@@ -23,166 +23,501 @@ import (
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
 )
 
-func serviceProfileAccessPointTypeConfigSch() map[string]*schema.Schema {
+func fabricServiceProfileSchema() map[string]*schema.Schema {
+	return map[string]*schema.Schema{
+		"href": {
+			Type:        schema.TypeString,
+			Computed:    true,
+			Description: "Service Profile URI response attribute",
+		},
+		"type": {
+			Type:        schema.TypeString,
+			Required:    true,
+			Description: "Service profile type - L2_PROFILE, L3_PROFILE, ECIA_PROFILE, ECMC_PROFILE",
+		},
+		"visibility": {
+			Type:        schema.TypeString,
+			Optional:    true,
+			Description: "Service profile visibility - PUBLIC, PRIVATE",
+		},
+		"name": {
+			Type:        schema.TypeString,
+			Required:    true,
+			Description: "Customer-assigned service profile name",
+		},
+		"uuid": {
+			Type:        schema.TypeString,
+			Computed:    true,
+			Description: "Equinix assigned service profile identifier",
+		},
+		"description": {
+			Type:        schema.TypeString,
+			Required:    true,
+			Description: "User-provided service description",
+		},
+		"notifications": {
+			Type:        schema.TypeList,
+			Optional:    true,
+			Description: "Preferences for notifications on connection configuration or status changes",
+			Elem: &schema.Resource{
+				Schema: equinix_fabric_schema.NotificationSch(),
+			},
+		},
+		"access_point_type_configs": {
+			Type:        schema.TypeList,
+			Optional:    true,
+			Description: "Access point config information",
+			Elem: &schema.Resource{
+				Schema: createSPAccessPointTypeConfigSch(),
+			},
+		},
+		"custom_fields": {
+			Type:        schema.TypeList,
+			Optional:    true,
+			Description: "Custom Fields",
+			Elem: &schema.Resource{
+				Schema: createCustomFieldSch(),
+			},
+		},
+		"marketing_info": {
+			Type:        schema.TypeSet,
+			Optional:    true,
+			Description: "Marketing Info",
+			MaxItems:    1,
+			Elem: &schema.Resource{
+				Schema: createMarketingInfoSch(),
+			},
+		},
+		"ports": {
+			Type:        schema.TypeList,
+			Optional:    true,
+			Description: "Ports",
+			Elem: &schema.Resource{
+				Schema: createServiceProfileAccessPointColo(),
+			},
+		},
+		"virtual_devices": {
+			Type:        schema.TypeList,
+			Optional:    true,
+			Description: "Virtual Devices",
+			Elem: &schema.Resource{
+				Schema: createServiceProfileAccessPointVd(),
+			},
+		},
+		"allowed_emails": {
+			Type:        schema.TypeList,
+			Optional:    true,
+			Description: "Array of contact emails",
+			Elem: &schema.Schema{
+				Type: schema.TypeString,
+			},
+		},
+		"tags": {
+			Type:        schema.TypeList,
+			Description: "Tags attached to the connection",
+			Optional:    true,
+			Elem:        &schema.Schema{Type: schema.TypeString},
+		},
+		"metros": {
+			Type:        schema.TypeList,
+			Optional:    true,
+			Description: "Access point config information",
+			Elem: &schema.Resource{
+				Schema: createServiceMetroSch(),
+			},
+		},
+		"self_profile": {
+			Type:        schema.TypeBool,
+			Optional:    true,
+			Description: "Self Profile indicating if the profile is created for customer's  self use",
+		},
+		"state": {
+			Type:        schema.TypeString,
+			Optional:    true,
+			Description: "Service profile state - ACTIVE, PENDING_APPROVAL, DELETED, REJECTED",
+		},
+		"account": {
+			Type:        schema.TypeSet,
+			Computed:    true,
+			Description: "Service Profile Owner Account Information",
+			MaxItems:    1,
+			Elem: &schema.Resource{
+				Schema: equinix_fabric_schema.AccountSch(),
+			},
+		},
+		"project": {
+			Type:        schema.TypeSet,
+			Optional:    true,
+			Description: "Project information",
+			MaxItems:    1,
+			Elem: &schema.Resource{
+				Schema: equinix_fabric_schema.ProjectSch(),
+			},
+		},
+		"change_log": {
+			Type:        schema.TypeSet,
+			Computed:    true,
+			Description: "Captures connection lifecycle change information",
+			Elem: &schema.Resource{
+				Schema: equinix_fabric_schema.ChangeLogSch(),
+			},
+		},
+	}
+}
+
+func createCustomFieldSch() map[string]*schema.Schema {
+	return map[string]*schema.Schema{
+		"label": {
+			Type:        schema.TypeString,
+			Required:    true,
+			Description: "Label",
+		},
+		"description": {
+			Type:        schema.TypeString,
+			Optional:    true,
+			Description: "Description",
+		},
+		"required": {
+			Type:        schema.TypeBool,
+			Required:    true,
+			Description: "Required field",
+		},
+		"data_type": {
+			Type:        schema.TypeString,
+			Required:    true,
+			Description: "Data type",
+		},
+		"options": {
+			Type:        schema.TypeList,
+			Optional:    true,
+			Description: "Options",
+			Elem: &schema.Schema{
+				Type: schema.TypeString,
+			},
+		},
+		"capture_in_email": {
+			Type:        schema.TypeBool,
+			Optional:    true,
+			Description: "Required field",
+		},
+	}
+}
+
+func createMarketingInfoSch() map[string]*schema.Schema {
+	return map[string]*schema.Schema{
+		"logo": {
+			Type:        schema.TypeString,
+			Optional:    true,
+			Description: "Logo",
+		},
+		"promotion": {
+			Type:        schema.TypeBool,
+			Optional:    true,
+			Description: "Promotion",
+		},
+		"process_step": {
+			Type:        schema.TypeList,
+			Optional:    true,
+			Description: "Process Step",
+			Elem: &schema.Resource{
+				Schema: createProcessStepSch(),
+			},
+		},
+	}
+}
+
+func createProcessStepSch() map[string]*schema.Schema {
+	return map[string]*schema.Schema{
+		"title": {
+			Type:        schema.TypeString,
+			Optional:    true,
+			Description: "Title",
+		},
+		"sub_title": {
+			Type:        schema.TypeString,
+			Optional:    true,
+			Description: "Sub Title",
+		},
+		"description": {
+			Type:        schema.TypeString,
+			Optional:    true,
+			Description: "Description",
+		},
+	}
+}
+
+func createServiceProfileAccessPointColo() map[string]*schema.Schema {
 	return map[string]*schema.Schema{
 		"type": {
 			Type:        schema.TypeString,
-			Computed:    true,
+			Required:    true,
+			Description: "Colo/Port Type",
+		},
+		"uuid": {
+			Type:        schema.TypeString,
+			Required:    true,
+			Description: "Colo/Port Uuid",
+		},
+		"location": {
+			Type:        schema.TypeSet,
+			Optional:    true,
+			Description: "Colo/Port Location",
+			MaxItems:    1,
+			Elem: &schema.Resource{
+				Schema: equinix_fabric_schema.LocationSch(),
+			},
+		},
+		"seller_region": {
+			Type:        schema.TypeString,
+			Optional:    true,
+			Description: "Seller Region",
+		},
+		"seller_region_description": {
+			Type:        schema.TypeString,
+			Optional:    true,
+			Description: "Seller Region details",
+		},
+		"cross_connect_id": {
+			Type:        schema.TypeString,
+			Optional:    true,
+			Description: "Cross Connect Id",
+		},
+	}
+}
+
+func createServiceProfileAccessPointVd() map[string]*schema.Schema {
+	return map[string]*schema.Schema{
+		"type": {
+			Type:        schema.TypeString,
+			Required:    true,
+			Description: "Virtual Device Type",
+		},
+		"uuid": {
+			Type:        schema.TypeString,
+			Required:    true,
+			Description: "Virtual Device Uuid",
+		},
+		"location": {
+			Type:        schema.TypeSet,
+			Optional:    true,
+			Description: "Device Location",
+			MaxItems:    1,
+			Elem: &schema.Resource{
+				Schema: equinix_fabric_schema.LocationSch(),
+			},
+		},
+		"interface_uuid": {
+			Type:        schema.TypeString,
+			Optional:    true,
+			Description: "Device Interface Uuid",
+		},
+	}
+}
+
+func createServiceMetroSch() map[string]*schema.Schema {
+	return map[string]*schema.Schema{
+		"code": {
+			Type:        schema.TypeString,
+			Optional:    true,
+			Description: "Metro Code - Example SV",
+		},
+		"name": {
+			Type:        schema.TypeString,
+			Optional:    true,
+			Description: "Metro Name",
+		},
+		"ibxs": {
+			Type:        schema.TypeList,
+			Optional:    true,
+			Description: "IBX- Equinix International Business Exchange list",
+			Elem: &schema.Schema{
+				Type: schema.TypeString,
+			},
+		},
+		"in_trail": {
+			Type:        schema.TypeBool,
+			Optional:    true,
+			Description: "In Trail",
+		},
+		"display_name": {
+			Type:        schema.TypeString,
+			Optional:    true,
+			Description: "Display Name",
+		},
+		"seller_regions": {
+			Type:        schema.TypeMap,
+			Optional:    true,
+			Description: "Seller Regions",
+			Elem: &schema.Schema{
+				Type: schema.TypeString,
+			},
+		},
+	}
+}
+
+func createSPAccessPointTypeConfigSch() map[string]*schema.Schema {
+	return map[string]*schema.Schema{
+		"type": {
+			Type:        schema.TypeString,
+			Required:    true,
 			Description: "Type of access point type config - VD, COLO",
 		},
 		"uuid": {
 			Type:        schema.TypeString,
 			Computed:    true,
-			Description: "Equinix-assigned access point type config identifier",
+			Description: "Colo/Port Uuid",
 		},
 		"connection_redundancy_required": {
 			Type:        schema.TypeBool,
-			Computed:    true,
+			Optional:    true,
 			Description: "Mandate redundant connections",
 		},
 		"allow_bandwidth_auto_approval": {
 			Type:        schema.TypeBool,
-			Computed:    true,
+			Optional:    true,
 			Description: "Setting to enable or disable the ability of the buyer to change connection bandwidth without approval of the seller",
 		},
 		"allow_remote_connections": {
 			Type:        schema.TypeBool,
-			Computed:    true,
+			Optional:    true,
 			Description: "Setting to allow or prohibit remote connections to the service profile",
 		},
 		"allow_bandwidth_upgrade": {
 			Type:        schema.TypeBool,
-			Computed:    true,
+			Optional:    true,
 			Description: "Availability of a bandwidth upgrade. The default is false",
 		},
 		"connection_label": {
 			Type:        schema.TypeString,
-			Computed:    true,
-			Description: "An alternative term for connection in case it's different in your organization.",
+			Optional:    true,
+			Description: "Custom name for Connection",
 		},
 		"enable_auto_generate_service_key": {
 			Type:        schema.TypeBool,
-			Computed:    true,
-			Description: "Applicable to Verizon Software Defined Interconnect service profile.",
+			Optional:    true,
+			Description: "Enable auto generate service key",
 		},
 		"bandwidth_alert_threshold": {
 			Type:        schema.TypeFloat,
-			Computed:    true,
+			Optional:    true,
 			Description: "Percentage of port bandwidth at which an allocation alert is generated",
 		},
 		"allow_custom_bandwidth": {
 			Type:        schema.TypeBool,
-			Computed:    true,
+			Optional:    true,
 			Description: "Setting to enable or disable the ability of the buyer to customize the bandwidth",
 		},
 		"api_config": {
 			Type:        schema.TypeSet,
-			Computed:    true,
-			Description: "API configuration settings and preferences.",
+			Optional:    true,
+			Description: "Api configuration details",
+			MaxItems:    1,
 			Elem: &schema.Resource{
-				Schema: apiConfigSch(),
+				Schema: createApiConfigSch(),
 			},
 		},
 		"authentication_key": {
 			Type:        schema.TypeSet,
-			Computed:    true,
-			Description: "Authentication key configuration",
+			Optional:    true,
+			Description: "Authentication key details",
+			MaxItems:    1,
 			Elem: &schema.Resource{
-				Schema: authenticationKeySch(),
+				Schema: createAuthenticationKeySch(),
 			},
 		},
 		"link_protocol_config": {
 			Type:        schema.TypeSet,
-			Computed:    true,
-			Description: "Settings and preferences for the link protocol used at the access point.",
+			Optional:    true,
+			Description: "Link protocol configuration details",
+			MaxItems:    1,
 			Elem: &schema.Resource{
-				Schema: linkProtocolConfigSch(),
+				Schema: createLinkProtocolConfigSch(),
 			},
 		},
 		"supported_bandwidths": {
 			Type:        schema.TypeList,
-			Computed:    true,
-			Description: "Allowed connection bandwidths (Mbps).",
+			Optional:    true,
+			Description: "Supported bandwidths",
 			Elem:        &schema.Schema{Type: schema.TypeInt},
 		},
 	}
 }
 
-func apiConfigSch() map[string]*schema.Schema {
+func createApiConfigSch() map[string]*schema.Schema {
 	return map[string]*schema.Schema{
 		"api_available": {
 			Type:        schema.TypeBool,
-			Computed:    true,
+			Optional:    true,
 			Description: "Indicates if it's possible to establish connections based on the given service profile using the Equinix Fabric API.",
 		},
 		"equinix_managed_vlan": {
 			Type:        schema.TypeBool,
-			Computed:    true,
+			Optional:    true,
 			Description: "Setting indicating that the VLAN is managed by Equinix (true) or not (false)",
 		},
 		"allow_over_subscription": {
 			Type:        schema.TypeBool,
-			Computed:    true,
+			Optional:    true,
 			Description: "Setting showing that oversubscription support is available (true) or not (false). The default is false",
 		},
 		"over_subscription_limit": {
 			Type:        schema.TypeInt,
-			Computed:    true,
-			Description: "Port bandwidth multiplier that determines the total bandwidth that can be allocated to users creating connections to your services.",
+			Optional:    true,
+			Description: "Port bandwidth multiplier that determines the total bandwidth that can be allocated to users creating connections to your services. For example, a 10 Gbps port combined with an overSubscriptionLimit parameter value of 10 allows your subscribers to create connections with a total bandwidth of 100 Gbps.",
 		},
 		"bandwidth_from_api": {
 			Type:        schema.TypeBool,
-			Computed:    true,
+			Optional:    true,
 			Description: "Indicates if the connection bandwidth can be obtained directly from the cloud service provider.",
 		},
 		"integration_id": {
 			Type:        schema.TypeString,
-			Computed:    true,
+			Optional:    true,
 			Description: "A unique identifier issued during onboarding and used to integrate the customer's service profile with the Equinix Fabric API.",
 		},
 		"equinix_managed_port": {
 			Type:        schema.TypeBool,
-			Computed:    true,
+			Optional:    true,
 			Description: "Setting indicating that the port is managed by Equinix (true) or not (false)",
 		},
 	}
 }
 
-func authenticationKeySch() map[string]*schema.Schema {
+func createAuthenticationKeySch() map[string]*schema.Schema {
 	return map[string]*schema.Schema{
 		"required": {
 			Type:        schema.TypeBool,
-			Computed:    true,
+			Optional:    true,
 			Description: "Requirement to configure an authentication key.",
 		},
 		"label": {
 			Type:        schema.TypeString,
-			Computed:    true,
+			Optional:    true,
 			Description: "Name of the parameter that must be provided to authorize the connection.",
 		},
 		"description": {
 			Type:        schema.TypeString,
-			Computed:    true,
+			Optional:    true,
 			Description: "Description of authorization key",
 		},
 	}
 }
 
-func linkProtocolConfigSch() map[string]*schema.Schema {
+func createLinkProtocolConfigSch() map[string]*schema.Schema {
 	return map[string]*schema.Schema{
 		"encapsulation_strategy": {
 			Type:        schema.TypeString,
-			Computed:    true,
+			Optional:    true,
 			Description: "Additional tagging information required by the seller profile.",
 		},
 		"reuse_vlan_s_tag": {
 			Type:        schema.TypeBool,
-			Computed:    true,
+			Optional:    true,
 			Description: "Automatically accept subsequent DOT1Q to QINQ connections that use the same authentication key. These connections will have the same VLAN S-tag assigned as the initial connection.",
 		},
 		"encapsulation": {
 			Type:        schema.TypeString,
-			Computed:    true,
+			Optional:    true,
 			Description: "Data frames encapsulation standard.UNTAGGED - Untagged encapsulation for EPL connections. DOT1Q - DOT1Q encapsulation standard. QINQ - QINQ encapsulation standard.",
 		},
 	}
@@ -203,7 +538,7 @@ func resourceFabricServiceProfile() *schema.Resource {
 		Importer: &schema.ResourceImporter{
 			StateContext: schema.ImportStatePassthroughContext,
 		},
-		Schema:      createFabricServiceProfileSchema(),
+		Schema:      fabricServiceProfileSchema(),
 		Description: "Fabric V4 API compatible resource allows creation and management of Equinix Fabric Service Profile",
 	}
 }
@@ -431,28 +766,18 @@ func WaitAndCheckServiceProfileDeleted(uuid string, client *v4.APIClient, ctx co
 	return err
 }
 
-func setFabricServiceProfileMap(d *schema.ResourceData, sp v4.ServiceProfile) diag.Diagnostics {
+func setFabricServiceProfilesListMap(d *schema.ResourceData, spl v4.ServiceProfiles) diag.Diagnostics {
 	diags := diag.Diagnostics{}
+	mappedServiceProfiles := make([]map[string]interface{}, len(spl.Data))
+	if spl.Data != nil {
+		for index, serviceProfile := range spl.Data {
+			mappedServiceProfiles[index] = fabricServiceProfileMap(&serviceProfile)
+		}
+	} else {
+		mappedServiceProfiles = nil
+	}
 	err := equinix_schema.SetMap(d, map[string]interface{}{
-		"href":                      sp.Href,
-		"type":                      sp.Type_,
-		"name":                      sp.Name,
-		"uuid":                      sp.Uuid,
-		"description":               sp.Description,
-		"notifications":             equinix_fabric_schema.NotificationsToTerra(sp.Notifications),
-		"tags":                      tagsFabricSpToTerra(sp.Tags),
-		"visibility":                sp.Visibility,
-		"access_point_type_configs": accessPointTypeConfigToTerra(sp.AccessPointTypeConfigs),
-		"custom_fields":             customFieldFabricSpToTerra(sp.CustomFields),
-		"marketing_info":            marketingInfoMappingToTerra(sp.MarketingInfo),
-		"ports":                     accessPointColoFabricSpToTerra(sp.Ports),
-		"allowed_emails":            allowedEmailsFabricSpToTerra(sp.AllowedEmails),
-		"metros":                    serviceMetroFabricSpToTerra(sp.Metros),
-		"self_profile":              sp.SelfProfile,
-		"state":                     sp.State,
-		"account":                   serviceProfileAccountFabricSpToTerra(sp.Account),
-		"project":                   equinix_fabric_schema.ProjectToTerra(sp.Project),
-		"change_log":                allOfServiceProfileChangeLogToTerra(sp.ChangeLog),
+		"data": mappedServiceProfiles,
 	})
 	if err != nil {
 		return diag.FromErr(err)
@@ -460,15 +785,40 @@ func setFabricServiceProfileMap(d *schema.ResourceData, sp v4.ServiceProfile) di
 	return diags
 }
 
-func setFabricServiceProfilesListMap(d *schema.ResourceData, spl v4.ServiceProfiles) diag.Diagnostics {
+func setFabricServiceProfileMap(d *schema.ResourceData, sp v4.ServiceProfile) diag.Diagnostics {
 	diags := diag.Diagnostics{}
-	err := equinix_schema.SetMap(d, map[string]interface{}{
-		"data": fabricServiceProfilesListToTerra(spl),
-	})
+	err := equinix_schema.SetMap(d, fabricServiceProfileMap(&sp))
 	if err != nil {
 		return diag.FromErr(err)
 	}
 	return diags
+}
+
+func fabricServiceProfileMap(serviceProfile *v4.ServiceProfile) map[string]interface{} {
+	if serviceProfile == nil {
+		return nil
+	}
+	return map[string]interface{}{
+		"href":                      serviceProfile.Href,
+		"type":                      serviceProfile.Type_,
+		"name":                      serviceProfile.Name,
+		"uuid":                      serviceProfile.Uuid,
+		"description":               serviceProfile.Description,
+		"notifications":             equinix_fabric_schema.NotificationsToTerra(serviceProfile.Notifications),
+		"tags":                      tagsFabricSpToTerra(serviceProfile.Tags),
+		"visibility":                serviceProfile.Visibility,
+		"access_point_type_configs": accessPointTypeConfigToTerra(serviceProfile.AccessPointTypeConfigs),
+		"custom_fields":             customFieldFabricSpToTerra(serviceProfile.CustomFields),
+		"marketing_info":            marketingInfoMappingToTerra(serviceProfile.MarketingInfo),
+		"ports":                     accessPointColoFabricSpToTerra(serviceProfile.Ports),
+		"allowed_emails":            allowedEmailsFabricSpToTerra(serviceProfile.AllowedEmails),
+		"metros":                    serviceMetroFabricSpToTerra(serviceProfile.Metros),
+		"self_profile":              serviceProfile.SelfProfile,
+		"state":                     serviceProfile.State,
+		"account":                   equinix_fabric_schema.AccountToTerra(serviceProfile.Account),
+		"project":                   equinix_fabric_schema.ProjectToTerra(serviceProfile.Project),
+		"change_log":                equinix_fabric_schema.ChangeLogToTerra(serviceProfile.ChangeLog),
+	}
 }
 
 func resourceServiceProfilesSearchRequest(ctx context.Context, d *schema.ResourceData, meta interface{}) diag.Diagnostics {
@@ -512,4 +862,383 @@ func resourceServiceProfilesSearchRequest(ctx context.Context, d *schema.Resourc
 	}
 	d.SetId(serviceProfiles.Data[0].Uuid)
 	return setFabricServiceProfilesListMap(d, serviceProfiles)
+}
+
+func customFieldFabricSpToTerra(customFieldl []v4.CustomField) []map[string]interface{} {
+	if customFieldl == nil {
+		return nil
+	}
+	mappedCustomFieldl := make([]map[string]interface{}, len(customFieldl))
+	for index, customField := range customFieldl {
+		mappedCustomFieldl[index] = map[string]interface{}{
+			"label":       customField.Label,
+			"description": customField.Description,
+			"required":    customField.Required,
+			"data_type":   customField.DataType,
+			"options":     customField.Options,
+		}
+	}
+	return mappedCustomFieldl
+}
+
+func processStepFabricSpToTerra(processStepl []v4.ProcessStep) []map[string]interface{} {
+	if processStepl == nil {
+		return nil
+	}
+	mappedProcessStepl := make([]map[string]interface{}, len(processStepl))
+	for index, processStep := range processStepl {
+		mappedProcessStepl[index] = map[string]interface{}{
+			"title":       processStep.Title,
+			"sub_title":   processStep.SubTitle,
+			"description": processStep.Description,
+		}
+	}
+	return mappedProcessStepl
+}
+
+func marketingInfoMappingToTerra(mkinfo *v4.MarketingInfo) *schema.Set {
+	if mkinfo == nil {
+		return nil
+	}
+	mkinfos := []*v4.MarketingInfo{mkinfo}
+	mappedMkInfos := make([]interface{}, 0)
+	for _, mkinfo := range mkinfos {
+		mappedMkInfo := make(map[string]interface{})
+		mappedMkInfo["logo"] = mkinfo.Logo
+		mappedMkInfo["promotion"] = mkinfo.Promotion
+		processStepl := processStepFabricSpToTerra(mkinfo.ProcessSteps)
+		if processStepl != nil {
+			mappedMkInfo["process_step"] = processStepl
+		}
+		mappedMkInfos = append(mappedMkInfos, mappedMkInfo)
+	}
+	marketingInfoSet := schema.NewSet(
+		schema.HashResource(&schema.Resource{Schema: createMarketingInfoSch()}),
+		mappedMkInfos,
+	)
+	return marketingInfoSet
+}
+
+func accessPointColoFabricSpToTerra(accessPointColol []v4.ServiceProfileAccessPointColo) []map[string]interface{} {
+	if accessPointColol == nil {
+		return nil
+	}
+	mappedAccessPointColol := make([]map[string]interface{}, len(accessPointColol))
+	for index, accessPointColo := range accessPointColol {
+		mappedAccessPointColol[index] = map[string]interface{}{
+			"type":                      accessPointColo.Type_,
+			"uuid":                      accessPointColo.Uuid,
+			"location":                  equinix_fabric_schema.LocationToTerra(accessPointColo.Location),
+			"seller_region":             accessPointColo.SellerRegion,
+			"seller_region_description": accessPointColo.SellerRegionDescription,
+			"cross_connect_id":          accessPointColo.CrossConnectId,
+		}
+	}
+	return mappedAccessPointColol
+}
+
+func serviceMetroFabricSpToTerra(serviceMetrol []v4.ServiceMetro) []map[string]interface{} {
+	if serviceMetrol == nil {
+		return nil
+	}
+	mappedServiceMetrol := make([]map[string]interface{}, len(serviceMetrol))
+	for index, serviceMetro := range serviceMetrol {
+		mappedServiceMetrol[index] = map[string]interface{}{
+			"code":           serviceMetro.Code,
+			"name":           serviceMetro.Name,
+			"ibxs":           serviceMetro.Ibxs,
+			"in_trail":       serviceMetro.InTrail,
+			"display_name":   serviceMetro.DisplayName,
+			"seller_regions": serviceMetro.SellerRegions,
+		}
+	}
+	return mappedServiceMetrol
+}
+
+func tagsFabricSpToTerra(tagsl *[]string) []interface{} {
+	if tagsl == nil {
+		return nil
+	}
+	mappedTags := make([]interface{}, 0)
+	for _, tag := range *tagsl {
+		mappedTags = append(mappedTags, tag)
+	}
+	return mappedTags
+}
+
+func allowedEmailsFabricSpToTerra(allowedemails []string) []interface{} {
+	if allowedemails == nil {
+		return nil
+	}
+	mappedEmails := make([]interface{}, len(allowedemails))
+	for index, email := range allowedemails {
+		mappedEmails[index] = email
+	}
+	return mappedEmails
+}
+
+func accessPointTypeConfigsToFabric(schemaAccessPointTypeConfigs []interface{}) []v4.ServiceProfileAccessPointType {
+	if schemaAccessPointTypeConfigs == nil {
+		return []v4.ServiceProfileAccessPointType{}
+	}
+	var accessPointTypeConfigs []v4.ServiceProfileAccessPointType
+	for _, accessPoint := range schemaAccessPointTypeConfigs {
+		spType := v4.ServiceProfileAccessPointTypeEnum(accessPoint.(map[string]interface{})["type"].(string))
+		spConnectionRedundancyRequired := accessPoint.(map[string]interface{})["connection_redundancy_required"].(bool)
+		spAllowBandwidthAutoApproval := accessPoint.(map[string]interface{})["allow_bandwidth_auto_approval"].(bool)
+		spAllowRemoteConnections := accessPoint.(map[string]interface{})["allow_remote_connections"].(bool)
+		spConnectionLabel := accessPoint.(map[string]interface{})["connection_label"].(string)
+		spEnableAutoGenerateServiceKey := accessPoint.(map[string]interface{})["enable_auto_generate_service_key"].(bool)
+		spBandwidthAlertThreshold := accessPoint.(map[string]interface{})["bandwidth_alert_threshold"].(float64)
+		spAllowCustomBandwidth := accessPoint.(map[string]interface{})["allow_custom_bandwidth"].(bool)
+
+		var spApiConfig v4.ApiConfig
+		if accessPoint.(map[string]interface{})["api_config"] != nil {
+			apiConfig := accessPoint.(map[string]interface{})["api_config"].(interface{}).(*schema.Set).List()
+			spApiConfig = apiConfigToFabric(apiConfig)
+		}
+
+		var spAuthenticationKey v4.AuthenticationKey
+		if accessPoint.(map[string]interface{})["authentication_key"] != nil {
+			authenticationKey := accessPoint.(map[string]interface{})["authentication_key"].(interface{}).(*schema.Set).List()
+			spAuthenticationKey = authenticationKeyToFabric(authenticationKey)
+		}
+
+		supportedBandwidthsRaw := accessPoint.(map[string]interface{})["supported_bandwidths"].([]interface{})
+		spSupportedBandwidths := converters.ListToInt32List(supportedBandwidthsRaw)
+
+		accessPointTypeConfigs = append(accessPointTypeConfigs, v4.ServiceProfileAccessPointType{
+			Type_:                        &spType,
+			ConnectionRedundancyRequired: spConnectionRedundancyRequired,
+			AllowBandwidthAutoApproval:   spAllowBandwidthAutoApproval,
+			AllowRemoteConnections:       spAllowRemoteConnections,
+			ConnectionLabel:              spConnectionLabel,
+			EnableAutoGenerateServiceKey: spEnableAutoGenerateServiceKey,
+			BandwidthAlertThreshold:      spBandwidthAlertThreshold,
+			AllowCustomBandwidth:         spAllowCustomBandwidth,
+			ApiConfig:                    &spApiConfig,
+			AuthenticationKey:            &spAuthenticationKey,
+			SupportedBandwidths:          &spSupportedBandwidths,
+		})
+	}
+	return accessPointTypeConfigs
+}
+
+func apiConfigToFabric(apiConfigl []interface{}) v4.ApiConfig {
+	if apiConfigl == nil {
+		return v4.ApiConfig{}
+	}
+	var apiConfigs v4.ApiConfig
+	for _, apiCongig := range apiConfigl {
+		psApiAvailable := apiCongig.(map[string]interface{})["api_available"].(interface{}).(bool)
+		psEquinixManagedVlan := apiCongig.(map[string]interface{})["equinix_managed_vlan"].(interface{}).(bool)
+		psBandwidthFromApi := apiCongig.(map[string]interface{})["bandwidth_from_api"].(interface{}).(bool)
+		psIntegrationId := apiCongig.(map[string]interface{})["integration_id"].(interface{}).(string)
+		psEquinixManagedPort := apiCongig.(map[string]interface{})["equinix_managed_port"].(interface{}).(bool)
+		apiConfigs = v4.ApiConfig{
+			ApiAvailable:       psApiAvailable,
+			EquinixManagedVlan: psEquinixManagedVlan,
+			BandwidthFromApi:   psBandwidthFromApi,
+			IntegrationId:      psIntegrationId,
+			EquinixManagedPort: psEquinixManagedPort,
+		}
+	}
+	return apiConfigs
+}
+
+func authenticationKeyToFabric(authenticationKeyl []interface{}) v4.AuthenticationKey {
+	if authenticationKeyl == nil {
+		return v4.AuthenticationKey{}
+	}
+	var authenticationKeys v4.AuthenticationKey
+	for _, authKey := range authenticationKeyl {
+		psRequired := authKey.(map[string]interface{})["required"].(interface{}).(bool)
+		psLabel := authKey.(map[string]interface{})["label"].(interface{}).(string)
+		psDescription := authKey.(map[string]interface{})["description"].(interface{}).(string)
+		authenticationKeys = v4.AuthenticationKey{
+			Required:    psRequired,
+			Label:       psLabel,
+			Description: psDescription,
+		}
+	}
+	return authenticationKeys
+}
+
+func customFieldsToFabric(schemaCustomField []interface{}) []v4.CustomField {
+	if schemaCustomField == nil {
+		return []v4.CustomField{}
+	}
+	var customFields []v4.CustomField
+	for _, customField := range schemaCustomField {
+		cfLabel := customField.(map[string]interface{})["label"].(string)
+		cfDescription := customField.(map[string]interface{})["description"].(string)
+		cfRequired := customField.(map[string]interface{})["required"].(bool)
+		cfDataType := customField.(map[string]interface{})["data_type"].(string)
+		optionsRaw := customField.(map[string]interface{})["options"].([]interface{})
+		cfOptions := converters.IfArrToStringArr(optionsRaw)
+		cfCaptureInEmail := customField.(map[string]interface{})["capture_in_email"].(bool)
+		customFields = append(customFields, v4.CustomField{
+			Label:          cfLabel,
+			Description:    cfDescription,
+			Required:       cfRequired,
+			DataType:       cfDataType,
+			Options:        cfOptions,
+			CaptureInEmail: cfCaptureInEmail,
+		})
+	}
+	return customFields
+}
+
+func marketingInfoToFabric(schemaMarketingInfo []interface{}) v4.MarketingInfo {
+	if schemaMarketingInfo == nil {
+		return v4.MarketingInfo{}
+	}
+	marketingInfos := v4.MarketingInfo{}
+	for _, marketingInfo := range schemaMarketingInfo {
+		miLogo := marketingInfo.(map[string]interface{})["logo"].(string)
+		miPromotion := marketingInfo.(map[string]interface{})["promotion"].(bool)
+
+		var miProcessSteps []v4.ProcessStep
+		if marketingInfo.(map[string]interface{})["process_steps"] != nil {
+			processStepsList := marketingInfo.(map[string]interface{})["process_steps"].([]interface{})
+			miProcessSteps = processStepToFabric(processStepsList)
+		}
+
+		marketingInfos = v4.MarketingInfo{
+			Logo:         miLogo,
+			Promotion:    miPromotion,
+			ProcessSteps: miProcessSteps,
+		}
+	}
+	return marketingInfos
+}
+
+func processStepToFabric(processStepl []interface{}) []v4.ProcessStep {
+	if processStepl == nil {
+		return []v4.ProcessStep{}
+	}
+	var processSteps []v4.ProcessStep
+	for _, processStep := range processStepl {
+		psTitle := processStep.(map[string]interface{})["title"].(interface{}).(string)
+		psSubTitle := processStep.(map[string]interface{})["sub_title"].(interface{}).(string)
+		psDescription := processStep.(map[string]interface{})["description"].(interface{}).(string)
+		processSteps = append(processSteps, v4.ProcessStep{
+			Title:       psTitle,
+			SubTitle:    psSubTitle,
+			Description: psDescription,
+		})
+	}
+	return processSteps
+}
+
+func portsToFabric(schemaPorts []interface{}) []v4.ServiceProfileAccessPointColo {
+	if schemaPorts == nil {
+		return []v4.ServiceProfileAccessPointColo{}
+	}
+	var ports []v4.ServiceProfileAccessPointColo
+	for _, port := range schemaPorts {
+		pType := port.(map[string]interface{})["type"].(string)
+		pUuid := port.(map[string]interface{})["uuid"].(string)
+		locationList := port.(map[string]interface{})["location"].(interface{}).(*schema.Set).List()
+		pLocation := v4.SimplifiedLocation{}
+		if len(locationList) != 0 {
+			pLocation = equinix_fabric_schema.LocationToFabric(locationList)
+		}
+		pSellerRegion := port.(map[string]interface{})["seller_region"].(string)
+		pSellerRegionDescription := port.(map[string]interface{})["seller_region_description"].(string)
+		pCrossConnectId := port.(map[string]interface{})["cross_connect_id"].(string)
+		ports = append(ports, v4.ServiceProfileAccessPointColo{
+			Type_:                   pType,
+			Uuid:                    pUuid,
+			Location:                &pLocation,
+			SellerRegion:            pSellerRegion,
+			SellerRegionDescription: pSellerRegionDescription,
+			CrossConnectId:          pCrossConnectId,
+		})
+	}
+	return ports
+}
+
+func virtualDevicesToFabric(schemaVirtualDevices []interface{}) []v4.ServiceProfileAccessPointVd {
+	if schemaVirtualDevices == nil {
+		return []v4.ServiceProfileAccessPointVd{}
+	}
+	var virtualDevices []v4.ServiceProfileAccessPointVd
+	for _, virtualDevice := range schemaVirtualDevices {
+		vType := virtualDevice.(map[string]interface{})["type"].(string)
+		vUuid := virtualDevice.(map[string]interface{})["uuid"].(string)
+		locationList := virtualDevice.(map[string]interface{})["location"].(interface{}).(*schema.Set).List()
+		vLocation := v4.SimplifiedLocation{}
+		if len(locationList) != 0 {
+			vLocation = equinix_fabric_schema.LocationToFabric(locationList)
+		}
+		pInterfaceUuid := virtualDevice.(map[string]interface{})["interface_uuid"].(string)
+		virtualDevices = append(virtualDevices, v4.ServiceProfileAccessPointVd{
+			Type_:         vType,
+			Uuid:          vUuid,
+			Location:      &vLocation,
+			InterfaceUuid: pInterfaceUuid,
+		})
+	}
+	return virtualDevices
+}
+
+func metrosToFabric(schemaMetros []interface{}) []v4.ServiceMetro {
+	if schemaMetros == nil {
+		return []v4.ServiceMetro{}
+	}
+	var metros []v4.ServiceMetro
+	for _, metro := range schemaMetros {
+		mCode := metro.(map[string]interface{})["code"].(string)
+		mName := metro.(map[string]interface{})["name"].(string)
+		ibxsRaw := metro.(map[string]interface{})["ibxs"].([]interface{})
+		mIbxs := converters.IfArrToStringArr(ibxsRaw)
+		mInTrail := metro.(map[string]interface{})["in_trail"].(bool)
+		mDisplayName := metro.(map[string]interface{})["display_name"].(string)
+		mSellerRegions := metro.(map[string]interface{})["seller_regions"].(map[string]string)
+		metros = append(metros, v4.ServiceMetro{
+			Code:          mCode,
+			Name:          mName,
+			Ibxs:          mIbxs,
+			InTrail:       mInTrail,
+			DisplayName:   mDisplayName,
+			SellerRegions: mSellerRegions,
+		})
+	}
+	return metros
+}
+
+func serviceProfilesSearchFilterRequestToFabric(schemaServiceProfileFilterRequest []interface{}) v4.ServiceProfileSimpleExpression {
+	if schemaServiceProfileFilterRequest == nil {
+		return v4.ServiceProfileSimpleExpression{}
+	}
+	mappedFilter := v4.ServiceProfileSimpleExpression{}
+	for _, s := range schemaServiceProfileFilterRequest {
+		sProperty := s.(map[string]interface{})["property"].(string)
+		operator := s.(map[string]interface{})["operator"].(string)
+		valuesRaw := s.(map[string]interface{})["values"].([]interface{})
+		values := converters.IfArrToStringArr(valuesRaw)
+		mappedFilter = v4.ServiceProfileSimpleExpression{Property: sProperty, Operator: operator, Values: values}
+	}
+	return mappedFilter
+}
+
+func serviceProfilesSearchSortRequestToFabric(schemaServiceProfilesSearchSortRequest []interface{}) []v4.ServiceProfileSortCriteria {
+	if schemaServiceProfilesSearchSortRequest == nil {
+		return []v4.ServiceProfileSortCriteria{}
+	}
+	var spSortCriteria []v4.ServiceProfileSortCriteria
+	for _, sp := range schemaServiceProfilesSearchSortRequest {
+		serviceProfileSortCriteriaMap := sp.(map[string]interface{})
+		direction := serviceProfileSortCriteriaMap["direction"]
+		directionCont := v4.ServiceProfileSortDirection(direction.(string))
+		property := serviceProfileSortCriteriaMap["property"]
+		propertyCont := v4.ServiceProfileSortBy(property.(string))
+		spSortCriteria = append(spSortCriteria, v4.ServiceProfileSortCriteria{
+			Direction: &directionCont,
+			Property:  &propertyCont,
+		})
+
+	}
+	return spSortCriteria
 }
